feat: memory read, probe protocol setter

+++ inc/psprobe/base.h
@@ -1,4 +1,6 @@
 
 #pragma once
 
-typedef size_t psprobe_status;
+#include <stdint.h>
+
+typedef uint32_t psprobe_status;
